разграничение по пользователям

+++ WorkManager/WorkManager/Controllers/WorksController.cs
@@ -23,11 +23,40 @@ namespace WorkManager.Controllers
             this.context = context;
         }
 
-        // GET: api/Works
+        // GET: api/Works?offset=1&count=3&sort=priority&category=study
         [HttpGet]
-        public IEnumerable<Work> GetWorks()
+        public IEnumerable<Work> GetWorks([FromQuery] WorkFilter filter)
         {
-            return context.Works.Where(work => work.User.UserName == User.Identity.Name);
+            var query = context.Works.Where(work => BelongsToCurrentUser(work));
+
+            if (filter.Category != null)
+            {
+                query = query.Where(work => work.Category == filter.Category);
+            }
+
+            query = query
+                .Skip(filter.Offset ?? 0)
+                .Take(filter.Count ?? 0);
+
+            switch (filter.SortBy)
+            {
+                case "priority":
+                    return query.OrderBy(w => w.Priority);
+
+                case "duration":
+                    return query.OrderBy(w => w.Duration);
+
+                case "difficulty":
+                    return query.OrderBy(w => w.Difficulty);
+
+                default:
+                    return query;
+            }
+        }
+
+        private bool BelongsToCurrentUser(Work work)
+        {
+            return work.User.UserName == User.Identity.Name;
         }
 
         // GET: api/Works/5
@@ -46,12 +75,17 @@ namespace WorkManager.Controllers
                 return NotFound();
             }
 
+            if (!BelongsToCurrentUser(work))
+            {
+                return Forbid();
+            }
+
             return Ok(work);
         }
 
         // PUT: api/Works/5
         [HttpPut("{id}")]
-        public async Task<IActionResult> PutWork([FromRoute] int id, [FromBody] Work work)
+        public async Task<IActionResult> UpdateWork([FromRoute] int id, [FromBody] Work work)
         {
             if (!ModelState.IsValid)
             {
@@ -114,6 +148,11 @@ namespace WorkManager.Controllers
                 return NotFound();
             }
 
+            if (!BelongsToCurrentUser(work))
+            {
+                return Forbid();
+            }
+
             context.Works.Remove(work);
             await context.SaveChangesAsync();
 
@@ -124,5 +163,13 @@ namespace WorkManager.Controllers
         {
             return context.Works.Any(e => e.Id == id);
         }
+    }
+
+    public class WorkFilter
+    {
+        public int? Offset { get; set; }
+        public int? Count { get; set; }
+        public string SortBy { get; set; }
+        public string Category { get; set; }
     }
 }
